+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -25,6 +25,8 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         DiscouragedCommentLocationRule(),
         FunKeywordSpacingRule(),
         FunctionTypeReferenceSpacingRule(),
-        ModifierListSpacingRule()
+        ModifierListSpacingRule(),
+        CommentWrappingRule(),
+        KdocWrappingRule()
     )
 }